BAH-4911 | Refactor. Move SearchPatient to search extensions
