+++ apps/registration/public/locales/locale_en.json
@@ -1,5 +1,35 @@
 {
   "ADDITIONAL_IDENTIFIERS_HEADER_TITLE": "Additional Identifiers",
+  "COMMON_SEARCH_ADD_CRITERIA_BUTTON": "Add Search Criteria",
+  "COMMON_SEARCH_API_ERROR_MESSAGE": "An error occurred while fetching search results. Please try again.",
+  "COMMON_SEARCH_CONFIG_ERROR": "Unable to load Common Search configuration. Please try refreshing the page.",
+  "COMMON_SEARCH_CRITERIA_DATE_INPUT_FIELD_FROM": "Enter From",
+  "COMMON_SEARCH_CRITERIA_DATE_INPUT_FIELD_TO": "Enter To",
+  "COMMON_SEARCH_CRITERIA_NUMERIC_INPUT_FIELD_FROM": "Enter From",
+  "COMMON_SEARCH_CRITERIA_NUMERIC_INPUT_FIELD_TO": "Enter To",
+  "COMMON_SEARCH_CRITERIA_TAG_RANGE": "{{from}} to {{to}}",
+  "COMMON_SEARCH_CRITERION_LABEL": "Enter search value",
+  "COMMON_SEARCH_CRITERION_PATIENT_GIVEN_NAME_INVALID_FORMAT": "Please enter at least 3 characters",
+  "COMMON_SEARCH_CRITERION_REQUIRED": "Please select a search criterion",
+  "COMMON_SEARCH_EVALUATION_ERROR": "An error occurred while processing search results. Please try again.",
+  "COMMON_SEARCH_INVALID_EXPRESSION": "Invalid expression in search configuration. Please contact your administrator.",
+  "COMMON_SEARCH_INVALID_FORMAT": "Invalid format",
+  "COMMON_SEARCH_LABEL": "Common Search",
+  "COMMON_SEARCH_MODIFY_SEARCH_BUTTON": "Modify Search",
+  "COMMON_SEARCH_NO_LOCATION_ERROR": "No location selected. Please log in with a location to use search.",
+  "COMMON_SEARCH_NO_PRIVILEGE_ERROR": "No search context configured for the current user's role / privileges.",
+  "COMMON_SEARCH_NO_RESULTS": "No results found for the given search criteria.",
+  "COMMON_SEARCH_RANGE_ORDER_INVALID": "To value must be greater than or equal to From value",
+  "COMMON_SEARCH_RESULTS_TABLE_TITLE": "Search Results",
+  "COMMON_SEARCH_SEARCH_BUTTON": "Search",
+  "COMMON_SEARCH_SELECT_CONTEXT_PLACEHOLDER": "Select a context",
+  "COMMON_SEARCH_SELECT_CONTEXT_TITLE": "Search Context",
+  "COMMON_SEARCH_SELECT_LOCATION_LABEL": "Location",
+  "COMMON_SEARCH_SELECT_SEARCH_CRITERIA": "Search Criteria",
+  "COMMON_SEARCH_SELECT_SEARCH_CRITERIA_PLACEHOLDER": "Select a search criteria",
+  "COMMON_SEARCH_SUCCESS": "Search Complete",
+  "COMMON_SEARCH_SUCCESS_MESSAGE": "Your search criteria have been submitted.",
+  "COMMON_SEARCH_VALUE_REQUIRED": "Please enter a search value",
   "CREATE_PATIENT_ACCURACY": "Accuracy",
   "CREATE_PATIENT_ADDRESS_LINE_PLACEHOLDER": "Address Line",
   "CREATE_PATIENT_AGE": "Age",
@@ -92,9 +122,12 @@
   "ERROR_LOADING_PATIENT_PHOTO": "Error Loading Patient Photo",
   "ERROR_SAVING_PATIENT": "Error saving patient",
   "ERROR_UPDATING_PATIENT": "Error updating patient",
+  "EXTENSION_WIDGET_NOT_REGISTERED": "Extension widget not registered",
+  "HOME_LABEL": "Home",
   "LOADING_PATIENT_DETAILS": "Navigating to patient details...",
   "MONTHS_one": " month",
   "MONTHS_other": " months",
+  "NO_EXTENSIONS_CONFIGURED": "No extensions configured",
   "PATIENT_SAVED_SUCCESSFULLY": "Patient saved successfully",
   "PATIENT_UPDATED_SUCCESSFULLY": "Patient updated successfully",
   "PHONE_NUMBER": "Phone Number",
@@ -107,13 +140,16 @@
   "REGISTRATION_ADDRESS_POSTAL_CODE": "Pincode",
   "REGISTRATION_ADDRESS_STATE": "State",
   "REGISTRATION_INVALID_NUMBER_MAX_LENGTH": "Number cannot exceed 15 characters",
+  "REGISTRATION_LABEL": "Registration",
   "REGISTRATION_PATIENT_SEARCH_BREADCRUMB_HOME": "Home",
   "REGISTRATION_PATIENT_SEARCH_BUTTON_TITLE": "Search",
   "REGISTRATION_PATIENT_SEARCH_CUSTOM_ATTRIBUTE_EMPTY_MESSAGE": "No matching search results for {{searchTerm}}. Try adjusting your search parameters.",
   "REGISTRATION_PATIENT_SEARCH_EMPTY_MESSAGE": "Could not find patient with identifier/name '{{searchTerm}}'. Please verify the patient ID or name entered or create a new patient record.",
   "REGISTRATION_PATIENT_SEARCH_ERROR_MESSAGE": "An unexpected error occurred during search. Please try again later.",
+  "REGISTRATION_PATIENT_SEARCH_HEADER_ACTIONS": "Actions",
   "REGISTRATION_PATIENT_SEARCH_HEADER_AGE": "Age",
   "REGISTRATION_PATIENT_SEARCH_HEADER_ALTERNATE_PHONE_NUMBER": "Alternate Phone Number",
+  "REGISTRATION_PATIENT_SEARCH_HEADER_BIRTH_DATE": "Date of Birth",
   "REGISTRATION_PATIENT_SEARCH_HEADER_GENDER": "Gender",
   "REGISTRATION_PATIENT_SEARCH_HEADER_ID": "Patient ID",
   "REGISTRATION_PATIENT_SEARCH_HEADER_NAME": "Patient Name",
