+++ apps/registration/public/locales/locale_es.json
@@ -1,5 +1,35 @@
 {
   "ADDITIONAL_IDENTIFIERS_HEADER_TITLE": "Identificadores Adicionales",
+  "COMMON_SEARCH_ADD_CRITERIA_BUTTON": "Agregar Criterios de Búsqueda",
+  "COMMON_SEARCH_API_ERROR_MESSAGE": "Se produjo un error al obtener los resultados de búsqueda. Por favor, inténtelo de nuevo.",
+  "COMMON_SEARCH_CONFIG_ERROR": "No se pudo cargar la configuración de Búsqueda Común. Por favor intente refrescar la página.",
+  "COMMON_SEARCH_CRITERIA_DATE_INPUT_FIELD_FROM": "Ingrese Desde",
+  "COMMON_SEARCH_CRITERIA_DATE_INPUT_FIELD_TO": "Ingrese Hasta",
+  "COMMON_SEARCH_CRITERIA_NUMERIC_INPUT_FIELD_FROM": "Ingrese Desde",
+  "COMMON_SEARCH_CRITERIA_NUMERIC_INPUT_FIELD_TO": "Ingrese Hasta",
+  "COMMON_SEARCH_CRITERIA_TAG_RANGE": "{{from}} a {{to}}",
+  "COMMON_SEARCH_CRITERION_LABEL": "Ingrese valor de búsqueda",
+  "COMMON_SEARCH_CRITERION_PATIENT_GIVEN_NAME_INVALID_FORMAT": "Por favor ingrese al menos 3 caracteres",
+  "COMMON_SEARCH_CRITERION_REQUIRED": "Por favor seleccione un criterio de búsqueda",
+  "COMMON_SEARCH_EVALUATION_ERROR": "Se produjo un error al procesar los resultados de búsqueda. Por favor, inténtelo de nuevo.",
+  "COMMON_SEARCH_INVALID_EXPRESSION": "Expresión inválida en la configuración de búsqueda. Por favor contacte a su administrador.",
+  "COMMON_SEARCH_INVALID_FORMAT": "Formato no válido",
+  "COMMON_SEARCH_LABEL": "Búsqueda Común",
+  "COMMON_SEARCH_MODIFY_SEARCH_BUTTON": "Modificar Búsqueda",
+  "COMMON_SEARCH_NO_LOCATION_ERROR": "No se ha seleccionado ninguna ubicación. Inicie sesión con una ubicación para usar la búsqueda.",
+  "COMMON_SEARCH_NO_PRIVILEGE_ERROR": "Ningún contexto de búsqueda configurado para el rol / privilegios del usuario actual.",
+  "COMMON_SEARCH_NO_RESULTS": "No se encontraron resultados para los criterios de búsqueda dados.",
+  "COMMON_SEARCH_RANGE_ORDER_INVALID": "El valor final debe ser mayor o igual al valor inicial",
+  "COMMON_SEARCH_RESULTS_TABLE_TITLE": "Resultados de búsqueda",
+  "COMMON_SEARCH_SEARCH_BUTTON": "Buscar",
+  "COMMON_SEARCH_SELECT_CONTEXT_PLACEHOLDER": "Seleccionar un contexto",
+  "COMMON_SEARCH_SELECT_CONTEXT_TITLE": "Contexto de Búsqueda",
+  "COMMON_SEARCH_SELECT_LOCATION_LABEL": "Ubicación",
+  "COMMON_SEARCH_SELECT_SEARCH_CRITERIA": "Criterios de Búsqueda",
+  "COMMON_SEARCH_SELECT_SEARCH_CRITERIA_PLACEHOLDER": "Seleccionar un criterio de búsqueda",
+  "COMMON_SEARCH_SUCCESS": "Búsqueda Completa",
+  "COMMON_SEARCH_SUCCESS_MESSAGE": "Sus criterios de búsqueda han sido enviados.",
+  "COMMON_SEARCH_VALUE_REQUIRED": "Por favor ingrese un valor de búsqueda",
   "CREATE_PATIENT_ACCURACY": "Precisión",
   "CREATE_PATIENT_ADDRESS_LINE_PLACEHOLDER": "Línea de Dirección",
   "CREATE_PATIENT_AGE": "Edad",
@@ -92,9 +122,12 @@
   "ERROR_LOADING_PATIENT_PHOTO": "Error al cargar la foto del paciente",
   "ERROR_SAVING_PATIENT": "Error al guardar el paciente",
   "ERROR_UPDATING_PATIENT": "Error al actualizar el paciente",
+  "EXTENSION_WIDGET_NOT_REGISTERED": "Widget de extensión no registrado",
+  "HOME_LABEL": "Inicio",
+  "LOADING_PATIENT_DETAILS": "Navegando a los detalles del paciente...",
   "MONTHS_one": " mes",
   "MONTHS_other": " meses",
-  "LOADING_PATIENT_DETAILS": "Navegando a los detalles del paciente...",
+  "NO_EXTENSIONS_CONFIGURED": "No hay extensiones configuradas",
   "PATIENT_SAVED_SUCCESSFULLY": "Paciente guardado exitosamente",
   "PATIENT_UPDATED_SUCCESSFULLY": "Paciente actualizado exitosamente",
   "PHONE_NUMBER": "Número de Teléfono",
@@ -107,13 +140,16 @@
   "REGISTRATION_ADDRESS_POSTAL_CODE": "Código Postal",
   "REGISTRATION_ADDRESS_STATE": "Estado",
   "REGISTRATION_INVALID_NUMBER_MAX_LENGTH": "El número no puede exceder los 15 caracteres",
+  "REGISTRATION_LABEL": "Registro",
   "REGISTRATION_PATIENT_SEARCH_BREADCRUMB_HOME": "Hogar",
   "REGISTRATION_PATIENT_SEARCH_BUTTON_TITLE": "Buscar",
   "REGISTRATION_PATIENT_SEARCH_CUSTOM_ATTRIBUTE_EMPTY_MESSAGE": "No hay resultados de búsqueda para {{searchTerm}}. Intenta ajustar tus parámetros de búsqueda.",
   "REGISTRATION_PATIENT_SEARCH_EMPTY_MESSAGE": "No se pudo encontrar el paciente con el identificador/nombre {{searchTerm}}. Verifique el ID o el nombre del paciente ingresado o cree un nuevo registro.",
   "REGISTRATION_PATIENT_SEARCH_ERROR_MESSAGE": "Ocurrió un error inesperado durante la búsqueda. Por favor, inténtelo de nuevo más tarde.",
+  "REGISTRATION_PATIENT_SEARCH_HEADER_ACTIONS": "Acciones",
   "REGISTRATION_PATIENT_SEARCH_HEADER_AGE": "Edad",
   "REGISTRATION_PATIENT_SEARCH_HEADER_ALTERNATE_PHONE_NUMBER": "Número de Teléfono Alternativo",
+  "REGISTRATION_PATIENT_SEARCH_HEADER_BIRTH_DATE": "Fecha de Nacimiento",
   "REGISTRATION_PATIENT_SEARCH_HEADER_GENDER": "Género",
   "REGISTRATION_PATIENT_SEARCH_HEADER_ID": "ID del Paciente",
   "REGISTRATION_PATIENT_SEARCH_HEADER_NAME": "Nombre del Paciente",
