+++ packages/bahmni-widgets/src/search/searchPatient/SearchStrategy.interface.ts
