+++ packages/bahmni-widgets/src/search/searchPatient/__mocks__/mocks.ts
