+++ packages/bahmni-widgets/src/search/searchPatient/index.ts
@@ -1,0 +1,10 @@
+import { registerSearchWidget } from '../registry';
+import SearchPatient from './SearchPatient';
+
+registerSearchWidget({
+  key: 'defaultSearch',
+  component: SearchPatient,
+});
+
+export { default as SearchPatient } from './SearchPatient';
+export type { PatientSearchViewModel } from './utils';
