+++ packages/bahmni-widgets/src/search/searchPatient/strategies/NameOrIdSearchStrategy.ts
