+++ packages/bahmni-widgets/src/search/searchPatient/strategies/SearchStrategyRegistry.ts
