+++ packages/bahmni-widgets/src/search/searchPatient/styles/SearchPatient.module.scss
