--- packages/bahmni-widgets/src/searchPatient/index.ts
@@ -1,2 +1,0 @@
-export { default as SearchPatient } from './SearchPatient';
-export type { PatientSearchViewModel } from './utils';
